+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Updated [README.md]
 - Added [grep] `-l|-LeaveHeader` option
+- Rewrite [head], [tail] code
 
 ## [6.4.0] - 2023-10-16