+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated [README.md]
 - Added [grep] `-l|-LeaveHeader` option
 - Rewrite [head], [tail] code
+- Output Aliases when dot-sourcing function files
+- Restricted [Get-Histogram], [Plot-BarChart] from division by zero
 
 ## [6.4.0] - 2023-10-16